requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.2.0
cffi==1.14.6
chardet==3.0.4
Cython==0.29.21
discord.py==1.7.3
ffmpeg==1.4
idna==2.10
multidict==4.7.6
pycparser==2.20
PyNaCl==1.4.0
six==1.16.0
yarl==1.5.1
youtube-dl==2021.6.6
